Refactor viewmodel, rename file
